chore: Update resource ID column in alert.bicep and scheduledqueryruleAggregate.bicep
This commit updates the resource ID column in the alert.bicep and scheduledqueryruleAggregate.bicep files. In alert.bicep, the 'name' property is changed from '_ResourceId' to 'Computer'. In scheduledqueryruleAggregate.bicep, the 'resourceIdColumn' property is changed from 'empty(dimensions) ? '_ResourceId' : null' to 'empty(dimensions) ? null : '_ResourceId''. These changes ensure that the correct resource ID column is used for the respective files.