chore: Add get-serviceTag and get-paasquery functions
This commit adds two new functions, get-serviceTag and get-paasquery, to the common.psm1 module. The get-serviceTag function retrieves a specific tag based on the provided namespace and tag mappings. The get-paasquery function returns a query string that filters specific types of resources. These functions enhance the functionality of the module and improve code organization.